added support for opentelemetry
