+++ build.gradle
@@ -47,6 +47,11 @@ dependencies {
     implementation group: 'net.thisptr', name: 'jackson-jq', version: '1.0.0-preview.20210928'
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.12.5'
     implementation 'io.jaegertracing:jaeger-client:1.8.0'
+    implementation 'io.opentelemetry:opentelemetry-sdk:1.11.0'
+    implementation 'io.opentelemetry:opentelemetry-opentracing-shim:1.11.0-alpha'
+    implementation 'io.opentelemetry:opentelemetry-semconv:1.11.0-alpha'
+    implementation 'io.opentelemetry:opentelemetry-exporter-jaeger:1.11.0'
+    implementation 'io.opentelemetry:opentelemetry-extension-trace-propagators:1.11.0'
 
     testImplementation("io.temporal:temporal-testing:1.8.0")
     testImplementation ("io.temporal:temporal-testing") {
