+++ src/main/java/io/temporal/samples/tracing/Starter.java
@@ -17,14 +17,14 @@
  *  permissions and limitations under the License.
  */
 
-package io.temporal.samples.opentracing;
+package io.temporal.samples.tracing;
 
 import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowClientOptions;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.client.WorkflowStub;
 import io.temporal.opentracing.OpenTracingClientInterceptor;
-import io.temporal.samples.opentracing.workflow.TracingWorkflow;
+import io.temporal.samples.tracing.workflow.TracingWorkflow;
 import io.temporal.serviceclient.WorkflowServiceStubs;
 
 public class Starter {
@@ -32,11 +32,15 @@ public class Starter {
   public static final String TASK_QUEUE_NAME = "tracingTaskQueue";
 
   public static void main(String[] args) {
+    String type = "OpenTelemetry";
+    if (args.length == 1) {
+      type = args[0];
+    }
 
     // Set the OpenTracing client interceptor
     WorkflowClientOptions clientOptions =
         WorkflowClientOptions.newBuilder()
-            .setInterceptors(new OpenTracingClientInterceptor(JaegerUtils.getJaegerOptions()))
+            .setInterceptors(new OpenTracingClientInterceptor(JaegerUtils.getJaegerOptions(type)))
             .build();
     WorkflowClient client = WorkflowClient.newInstance(service, clientOptions);
